Siirra hyväksymisesityksen testeissä käytettyjä db-apufunktioita databaseUtilsiin ja päivitä importit
