+++ backend/integrationtest/HyvaksymisEsitys/esikatseleTiedostot.test.ts
@@ -1,9 +1,9 @@
 import sinon from "sinon";
 import { userService } from "../../src/user";
 import TEST_HYVAKSYMISESITYS, { TEST_HYVAKSYMISESITYS_FILES } from "./TEST_HYVAKSYMISESITYS";
-import { deleteYllapitoFiles, insertProjektiToDB, insertYllapitoFileToS3, removeProjektiFromDB } from "./util";
+import { deleteYllapitoFiles, insertYllapitoFileToS3 } from "./util";
 import { UserFixture } from "../../test/fixture/userFixture";
-import { setupLocalDatabase } from "../util/databaseUtil";
+import { insertProjektiToDB, removeProjektiFromDB, setupLocalDatabase } from "../util/databaseUtil";
 import { TEST_PROJEKTI, TEST_PROJEKTI_FILES } from "./TEST_PROJEKTI";
 import * as API from "hassu-common/graphql/apiModel";
 import { esikatseleHyvaksymisEsityksenTiedostot } from "../../src/HyvaksymisEsitys/actions";
