+++ backend/integrationtest/util/databaseUtil.ts
@@ -1,10 +1,13 @@
 import * as sinon from "sinon";
 import { SinonStub } from "sinon";
 
-import { DynamoDB } from "@aws-sdk/client-dynamodb";
-import { DynamoDBDocumentClient } from "@aws-sdk/lib-dynamodb";
+import { DeleteItemCommand, DynamoDB } from "@aws-sdk/client-dynamodb";
+import { DynamoDBDocumentClient, GetCommand, PutCommand } from "@aws-sdk/lib-dynamodb";
 import * as awsClient from "../../src/aws/client";
 import mocha from "mocha";
+import { DBProjekti } from "../../src/database/model";
+import { getDynamoDBDocumentClient } from "../../src/aws/client";
+import { config } from "../../src/config";
 
 const localDynamoDB = new DynamoDB({
   endpoint: "http://localhost:4566",
@@ -27,4 +30,37 @@ export function setupLocalDatabase(): void {
   mocha.beforeEach(() => {
     localDynamoDBDocumentClientStub.returns(localDocumentClient);
   });
+}
+
+export async function insertProjektiToDB<A extends Pick<DBProjekti, "oid">>(projekti: A) {
+  const params = new PutCommand({
+    TableName: config.projektiTableName,
+    Item: projekti,
+  });
+  await getDynamoDBDocumentClient().send(params);
+}
+
+export async function removeProjektiFromDB(oid: string) {
+  const params = new DeleteItemCommand({
+    TableName: config.projektiTableName,
+    Key: {
+      oid: {
+        S: oid,
+      },
+    },
+  });
+  await getDynamoDBDocumentClient().send(params);
+}
+
+export async function getProjektiFromDB(oid: string): Promise<any> {
+  const params = new GetCommand({
+    TableName: config.projektiTableName,
+    Key: { oid },
+    ConsistentRead: true,
+  });
+  const data = await getDynamoDBDocumentClient().send(params);
+  if (!data.Item) {
+    return;
+  }
+  return data.Item as any;
 }
